pyproject.toml
[tool.poetry]
name = "rnpo-web-tools"
version = "0.1.0"
description = ""
authors = ["Ramil Albakov <ralbakov10@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
cx-Oracle = "^8.3.0"
python-dotenv = "^0.20.0"
openpyxl = "^3.0.10"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
